requirements.txt
pydantic>=2.0
httpx>=0.24.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-vcr>=1.0.2
hypothesis>=6.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
stripe>=5.0.0
ShopifyAPI>=12.0.0
paypal-checkout-serversdk>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
